Round deterministic grad norms at the clipping boundary
Accumulate and reduce the L2 norm in FP64, then round the final scalar once to FP32. This keeps clipping coefficients identical across distributed-optimizer shard topologies without changing the real-rollout test contract.

+++ miles/backends/megatron_utils/model.py
@@ -122,7 +122,7 @@ def _get_grad_norm_fp64(
         torch.distributed.all_reduce(total_squared_norm, group=data_parallel_group)
     torch.distributed.all_reduce(total_squared_norm, group=grad_stats_parallel_group)
 
-    return total_squared_norm.sqrt().item()
+    return total_squared_norm.sqrt().float().item()
 
 
 def _run_stable_dp_shards(
